requirements.txt
requests==2.13.0
Flask==1.1.2
mysql-connector-python==8.0.23
apscheduler==3.7.0